requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
click-default-group==1.2.4
colorama==0.4.6
faiss-cpu==1.11.0
filelock==3.18.0
fsspec==2025.5.1
huggingface-hub==0.32.1
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
packaging==25.0
pillow==11.2.1
pluggy==1.6.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
sentence-transformers==4.1.0
setuptools==80.8.0
six==1.17.0
sqlite-fts4==1.0.3
sqlite-utils==3.38
sympy==1.14.0
tabulate==0.9.0
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
transformers==4.52.3
typing_extensions==4.13.2
urllib3==2.4.0
